pyproject.toml
[project]
name = "govsearch"
version = "0.1.0"
description = "Instead of using a chatbot approach like Govbot, we tried to solve the problem with a search-based approach."
dependencies = [
    "ariadne>=0.23.0",
    "pyright>=1.1.357",
    "uvicorn[standard]>=0.29.0",
    "starlette>=0.37.2",
    "gql-schema-codegen>=1.0.1",
    "websockets>=12.0",
    "openai>=1.16.0",
    "pyvespa>=0.39.0",
    "transformers>=4.39.3",
    "torch>=2.2.2",
    "onnx>=1.16.0",
    "onnxruntime>=1.17.1",
    "onnxcli>=0.2.1",
    "optimum[exporters,onnxruntime]>=1.18.0",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
    "async-caches>=0.3.0",
    "content-size-limit-asgi>=0.1.5",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pyright]
include = ["backend", "scripts"]
